fix: review uows caches (#811)
