+++ apps/backend/package.json
@@ -36,6 +36,7 @@
     "@apollo/server": "^4.10.0",
     "@apollo/subgraph": "^2.7.1",
     "@graphql-tools/utils": "^10.1.0",
+    "@isaacs/ttlcache": "^1.4.1",
     "@user-office-software/duo-localisation": "^1.2.0",
     "@user-office-software/duo-logger": "^2.1.1",
     "@user-office-software/duo-message-broker": "^1.6.0",
