+++ apps/backend/src/auth/StfcUserAuthorization.spec.ts
@@ -8,7 +8,7 @@ import { Instrument } from '../models/Instrument';
 import { StfcUserAuthorization } from './StfcUserAuthorization';
 
 jest.mock('../datasources/stfc/UOWSSoapInterface.ts');
-jest.mock('../utils/LRUCache');
+jest.mock('../utils/Cache');
 
 const userAuthorization = container.resolve(StfcUserAuthorization);
